refactor: move ArXiv helper functions into arxiv.ts domain handler
Why: formatArxivServiceMessage, buildArxivFallbackUrl, and
isLikelyArxivErrorBody were defined in src/index.ts but logically
belong in the arxiv domain handler.

Changes:
- Moved 3 ArXiv utility functions from src/index.ts to
  src/webfetch-handlers/domains/arxiv.ts
- Updated imports in src/index.ts accordingly
- Pure refactor: no behavior change
